ポインタをusizeにする

+++ src/cisc/internal.rs
@@ -34,11 +34,11 @@ pub fn u32_to_two_delta(val: u32) -> (i16, i16) {
 }
 
 #[inline(always)]
-pub fn positive_out_of_range(r: u8, pointer: isize) -> bool {
-    (r as i8 as i16 as u16 as isize) <= pointer
+pub fn positive_out_of_range(r: u8, pointer: usize) -> bool {
+    (r as i8 as i16 as u16 as isize) <= pointer as isize
 }
 
 #[inline(always)]
-pub fn negative_out_of_range(r: u8, pointer: isize) -> bool {
-    (r as i8 as i16 as u16 as isize) > pointer
+pub fn negative_out_of_range(r: u8, pointer: usize) -> bool {
+    (r as i8 as i16 as u16 as isize) > pointer as isize
 }
