+++ src/memory.rs
@@ -9,39 +9,39 @@ impl Memory {
 }
 
 impl Memory {
-    pub fn get(&self, index: isize) -> Result<u8, String> {
-        match self.0.get(index as usize) {
+    pub fn get(&self, index: usize) -> Result<u8, String> {
+        match self.0.get(index) {
             Some(val) => Ok(*val),
             None => Err(format!("Runtime Error: Out of range memory get, Address: {}", index)),
         }
     }
-    pub fn set(&mut self, index: isize, value: u8) -> Result<(), String> {
-        if self.0.len() <= index as usize {
+    pub fn set(&mut self, index: usize, value: u8) -> Result<(), String> {
+        if self.0.len() <= index {
             Err(format!("Runtime Error: Out of range memory set, Address: {}", index))
         } else {
             unsafe { // SAFETY: 直前に範囲を確認済み
-                *self.0.as_mut_ptr().add(index as usize) = value;
+                *self.0.as_mut_ptr().add(index) = value;
             }
             Ok(())
         }
     }
-    pub fn add(&mut self, index: isize, value: u8) -> Result<(), String> {
-        if self.0.len() <= index as usize {
+    pub fn add(&mut self, index: usize, value: u8) -> Result<(), String> {
+        if self.0.len() <= index {
             Err(format!("Runtime Error: Out of range memory add, Address: {}", index))
         } else {
             unsafe { // SAFETY: 直前に範囲を確認済み
-                let ptr = self.0.as_mut_ptr().add(index as usize);
+                let ptr = self.0.as_mut_ptr().add(index);
                 *ptr = (*ptr).wrapping_add(value);
             }
             Ok(())
         }
     }
-    pub fn sub(&mut self, index: isize, value: u8) -> Result<(), String> {
-        if self.0.len() <= index as usize {
+    pub fn sub(&mut self, index: usize, value: u8) -> Result<(), String> {
+        if self.0.len() <= index {
             Err(format!("Runtime Error: Out of range memory sub, Address: {}", index))
         } else {
             unsafe { // SAFETY: 直前に範囲を確認済み
-                let ptr = self.0.as_mut_ptr().add(index as usize);
+                let ptr = self.0.as_mut_ptr().add(index);
                 *ptr = (*ptr).wrapping_sub(value);
             }
             Ok(())
@@ -50,7 +50,7 @@ impl Memory {
     #[inline(always)]
     pub unsafe fn get_unchecked(&self, index: usize) -> u8 {
         #[cfg(feature = "debug")] {
-            if self.0.len() <= index as usize {
+            if self.0.len() <= index {
                 panic!("[UNSAFE] Runtime Error: Out of range memory get, Address: {} ", index);
             }
         }
@@ -59,7 +59,7 @@ impl Memory {
     #[inline(always)]
     pub unsafe fn set_unchecked(&mut self, index: usize, value: u8) {
         #[cfg(feature = "debug")] {
-            if self.0.len() <= index as usize {
+            if self.0.len() <= index {
                 panic!("[UNSAFE] Runtime Error: Out of range memory set, Address: {} ", index);
             }
         }
@@ -68,7 +68,7 @@ impl Memory {
     #[inline(always)]
     pub unsafe fn add_unchecked(&mut self, index: usize, value: u8) {
         #[cfg(feature = "debug")] {
-            if self.0.len() <= index as usize {
+            if self.0.len() <= index {
                 panic!("[UNSAFE] Runtime Error: Out of range memory add, Address: {} ", index);
             }
         }
@@ -78,7 +78,7 @@ impl Memory {
     #[inline(always)]
     pub unsafe fn sub_unchecked(&mut self, index: usize, value: u8) {
         #[cfg(feature = "debug")] {
-            if self.0.len() <= index as usize {
+            if self.0.len() <= index {
                 panic!("[UNSAFE] Runtime Error: Out of range memory sub, Address: {} ", index);
             }
         }
