+++ src/risc/interpret.rs
@@ -7,7 +7,7 @@ pub fn run(insts: &[Bytecode], memory: &mut Memory, ocm: &mut OperationCountMap)
     let mut stdin = stdin().lock();
     let mut stdin_buf: [u8; 1] = [0];
     let mut pc: usize = 0;
-    let mut pointer: isize = 0;
+    let mut pointer: usize = 0;
     let mut mul_val: u8 = 0;
     
     loop {
@@ -19,29 +19,29 @@ pub fn run(insts: &[Bytecode], memory: &mut Memory, ocm: &mut OperationCountMap)
         
         match bytecode.opcode {
             OpCode::Breakpoint => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 eprintln!("PC: {}, PTR: {}", pc, pointer);
             }
 
             OpCode::Add => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 memory.add(pointer, bytecode.val)?;
             }
             OpCode::Set => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 memory.set(pointer, bytecode.val)?;
             }
 
             OpCode::Shift => {
-                pointer += bytecode.delta as isize;
-                let step = bytecode.addr as i32 as isize;
+                pointer += bytecode.delta as isize as usize;
+                let step = bytecode.addr as i32 as isize as usize;
                 while memory.get(pointer)? != 0 {
                     pointer += step;
                 }
             }
 
             OpCode::MulStart => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 let val = memory.get(pointer)?;
                 if val == 0 {
                     pc = bytecode.addr as usize;
@@ -52,28 +52,28 @@ pub fn run(insts: &[Bytecode], memory: &mut Memory, ocm: &mut OperationCountMap)
                 }
             }
             OpCode::Mul => {
-                memory.add(pointer + bytecode.delta as isize, mul_val.wrapping_mul(bytecode.val))?;
+                memory.add(pointer + bytecode.delta as isize as usize, mul_val.wrapping_mul(bytecode.val))?;
             }
 
             OpCode::SingleMoveAdd => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 let v = memory.get(pointer)?;
                 if v != 0 {
                     memory.set(pointer, 0)?;
-                    memory.add(pointer + (bytecode.addr as i32 as isize), v)?;
+                    memory.add(pointer + (bytecode.addr as i32 as isize as usize), v)?;
                 }
             }
             OpCode::SingleMoveSub => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 let v = memory.get(pointer)?;
                 if v != 0 {
                     memory.set(pointer, 0)?;
-                    memory.sub(pointer + (bytecode.addr as i32 as isize), v)?;
+                    memory.sub(pointer + (bytecode.addr as i32 as isize as usize), v)?;
                 }
             }
 
             OpCode::MoveStart => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 let val = memory.get(pointer)?;
                 if val == 0 {
                     pc = bytecode.addr as usize;
@@ -84,33 +84,33 @@ pub fn run(insts: &[Bytecode], memory: &mut Memory, ocm: &mut OperationCountMap)
                 }
             }
             OpCode::MoveAdd => {
-                memory.add(pointer + bytecode.delta as isize, mul_val)?;
+                memory.add(pointer + bytecode.delta as isize as usize, mul_val)?;
             }
             OpCode::MoveSub => {
-                memory.sub(pointer + bytecode.delta as isize, mul_val)?;
+                memory.sub(pointer + bytecode.delta as isize as usize, mul_val)?;
             }
 
             OpCode::In => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 match stdin.read_exact(&mut stdin_buf) {
                     Ok(_) => memory.set(pointer, stdin_buf[0])?,
                     Err(_) => memory.set(pointer, 0)?,
                 }
             }
             OpCode::Out => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 stdout.write(&[memory.get(pointer)?]).map_err(|_| "Runtime Error: Failed to print")?;
             }
 
             OpCode::JmpIfZero => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 if memory.get(pointer)? == 0 {
                     pc = bytecode.addr as usize;
                     continue;
                 }
             }
             OpCode::JmpIfNotZero => {
-                pointer += bytecode.delta as isize;
+                pointer += bytecode.delta as isize as usize;
                 if memory.get(pointer)? != 0 {
                     pc = bytecode.addr as usize;
                     continue;
